Remove doubling of memory in the Clear method
Previous implementation of the Clear method has allocated new ringbuffers.
Memory footprint of the process has been doubled so as the old memory blocks
stayed in the memory until the garbage collector cleaned them.

This change In this commit the the beahvior has been changed: just internal
pointers in the ringbuffers are reset now and the old memory blocks are reused.